requirements.txt
config==0.5.1
loguru==0.5.3
python-dotenv==0.19.2
requests==2.26.0
